requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
httpx>=0.24.0,<0.25.0
supabase==2.0.3


